+++ backend/src/hatchling/build.py
@@ -3,7 +3,7 @@ import os
 
 def get_requires_for_build_sdist(config_settings=None):
     """
-    https://www.python.org/dev/peps/pep-0517/#get-requires-for-build-sdist
+    https://peps.python.org/pep-0517/#get-requires-for-build-sdist
     """
     from .builders.sdist import SdistBuilder
 
@@ -13,7 +13,7 @@ def get_requires_for_build_sdist(config_settings=None):
 
 def build_sdist(sdist_directory, config_settings=None):
     """
-    https://www.python.org/dev/peps/pep-0517/#build-sdist
+    https://peps.python.org/pep-0517/#build-sdist
     """
     from .builders.sdist import SdistBuilder
 
@@ -23,7 +23,7 @@ def build_sdist(sdist_directory, config_settings=None):
 
 def get_requires_for_build_wheel(config_settings=None):
     """
-    https://www.python.org/dev/peps/pep-0517/#get-requires-for-build-wheel
+    https://peps.python.org/pep-0517/#get-requires-for-build-wheel
     """
     from .builders.wheel import WheelBuilder
 
@@ -33,7 +33,7 @@ def get_requires_for_build_wheel(config_settings=None):
 
 def build_wheel(wheel_directory, config_settings=None, metadata_directory=None):
     """
-    https://www.python.org/dev/peps/pep-0517/#build-wheel
+    https://peps.python.org/pep-0517/#build-wheel
     """
     from .builders.wheel import WheelBuilder
 
@@ -43,7 +43,7 @@ def build_wheel(wheel_directory, config_settings=None, metadata_directory=None):
 
 def get_requires_for_build_editable(config_settings=None):
     """
-    https://www.python.org/dev/peps/pep-0660/#get-requires-for-build-editable
+    https://peps.python.org/pep-0660/#get-requires-for-build-editable
     """
     from .builders.wheel import WheelBuilder
 
@@ -53,7 +53,7 @@ def get_requires_for_build_editable(config_settings=None):
 
 def build_editable(wheel_directory, config_settings=None, metadata_directory=None):
     """
-    https://www.python.org/dev/peps/pep-0660/#build-editable
+    https://peps.python.org/pep-0660/#build-editable
     """
     from .builders.wheel import WheelBuilder